requirements.txt
openai
fastapi
uvicorn[standard]
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pydantic